Re-org unit tests: 1 hpp file per tested class
